@@ -9,82 +9,83 @@
     "scripts": {
         "dev": "gulp",
         "build": "gulp prod --production",
-        "build views": "gulp views --production",
-        "build styles": "gulp styles --production",
-        "build scripts": "gulp scripts --production",
-        "build images": "gulp images --production",
-        "build webp": "gulp webp --production",
-        "build sprites": "gulp sprites --production",
-        "build fonts": "gulp fonts --production",
-        "build favicons": "gulp favicons --production",
-        "build gzip": "gulp gzip --production",
+        "build:views": "gulp views --production",
+        "build:styles": "gulp styles --production",
+        "build:scripts": "gulp scripts --production",
+        "build:images": "gulp images --production",
+        "build:webp": "gulp webp --production",
+        "build:sprites": "gulp sprites --production",
+        "build:fonts": "gulp fonts --production",
+        "build:favicons": "gulp favicons --production",
+        "build:gzip": "gulp gzip --production",
         "lint:style": "stylelint --color \"src/**/*.scss\"",
         "lint:style --fix": "stylelint --fix --color \"src/**/*.scss\"",
         "lint:script": "eslint --color ./src/blocks",
         "lint:script --fix": "eslint --fix --color ./src/blocks"
     },
     "devDependencies": {
-        "@babel/core": "^7.4.3",
-        "@babel/preset-env": "^7.4.3",
-        "@babel/register": "^7.4.0",
+        "@babel/core": "^7.8.3",
+        "@babel/preset-env": "^7.8.3",
+        "@babel/register": "^7.8.3",
+        "autoprefixer": "^9.7.4",
         "babel-eslint": "^10.0.2",
         "babel-loader": "^8.0.5",
         "bem-tools-core": "bem-tools/bem-tools-core",
         "bem-tools-create": "bem-tools/bem-tools-create",
         "browser-sync": "^2.23.6",
-        "csscomb": "^4.3.0",
+        "css-mqpacker": "^7.0.0",
+        "cssnano": "^4.1.10",
         "cz-conventional-changelog": "^3.0.2",
-        "del": "^4.1.1",
-        "eslint": "^6.0.0",
+        "del": "^5.1.0",
+        "eslint": "^6.8.0",
         "eslint-config-airbnb": "^18.0.1",
-        "eslint-config-prettier": "^6.5.0",
-        "eslint-plugin-import": "^2.18.2",
-        "eslint-plugin-prettier": "^3.1.1",
+        "eslint-config-prettier": "^6.9.0",
+        "eslint-plugin-import": "^2.20.0",
+        "eslint-plugin-prettier": "^3.1.2",
         "gulp": "^4.0.0",
-        "gulp-autoprefixer": "^6.1.0",
         "gulp-cheerio": "^1.0.0",
-        "gulp-clean-css": "^4.2.0",
         "gulp-debug": "^4.0.0",
-        "gulp-favicons": "^2.3.1",
-        "gulp-group-css-media-queries": "^1.2.2",
-        "gulp-if": "^2.0.2",
-        "gulp-imagemin": "^6.0.0",
+        "gulp-favicons": "^2.4.0",
+        "gulp-if": "^3.0.0",
+        "gulp-imagemin": "^7.1.0",
         "gulp-plumber": "^1.2.0",
+        "gulp-postcss": "^8.0.0",
         "gulp-pug": "^4.0.1",
-        "gulp-rename": "^1.2.2",
+        "gulp-rename": "^2.0.0",
         "gulp-replace": "^1.0.0",
         "gulp-sass": "^4.0.2",
-        "gulp-sourcemaps": "^2.6.4",
+        "gulp-sourcemaps": "^2.6.5",
         "gulp-svg-sprite": "^1.5.0",
         "gulp-svgmin": "^2.2.0",
+        "gulp-typograf": "^4.0.2",
         "gulp-watch": "^5.0.0",
         "gulp-webp": "^4.0.1",
-        "husky": "^3.0.9",
+        "husky": "^4.0.10",
         "imagemin-giflossy": "^5.1.10",
         "imagemin-jpeg-recompress": "^6.0.0",
         "imagemin-mozjpeg": "^8.0.0",
         "imagemin-pngquant": "^8.0.0",
         "imagemin-webp": "^5.0.0",
         "imagemin-zopfli": "^6.0.0",
-        "lint-staged": "^9.4.3",
+        "lint-staged": "^10.0.1",
         "prettier": "^1.19.1",
         "require-dir": "^1.2.0",
         "smart-grid": "^2.1.2",
-        "stylelint": "^10.0.1",
-        "stylelint-config-prettier": "^6.0.0",
-        "stylelint-config-recommended-scss": "^4.0.0",
-        "stylelint-order": "^3.1.1",
+        "stylelint": "^13.0.0",
+        "stylelint-config-recommended-scss": "^4.1.0",
+        "stylelint-config-standard": "^19.0.0",
+        "stylelint-order": "^4.0.0",
         "stylelint-scss": "^3.12.1",
         "stylelint-selector-bem-pattern": "^2.1.0",
-        "webpack": "^4.35.0",
+        "webpack": "^4.41.5",
         "webpack-stream": "^5.2.1",
-        "yargs": "^15.0.2"
+        "yargs": "^15.1.0"
     },
     "dependencies": {
         "animejs": "^3.1.0",
-        "disable-scroll": "^0.5.0",
-        "perfect-scrollbar": "^1.4.0",
-        "swiper": "^5.2.1"
+        "disable-scroll": "^0.5.1",
+        "perfect-scrollbar": "^1.5.0",
+        "swiper": "^5.3.0"
     },
     "husky": {
         "hooks": {
