@@ -5,9 +5,10 @@ import gulp from "gulp";
 import gulpif from "gulp-if";
 import rename from "gulp-rename";
 import sass from "gulp-sass";
-import mincss from "gulp-clean-css";
-import groupmedia from "gulp-group-css-media-queries";
-import autoprefixer from "gulp-autoprefixer";
+import postcss from "gulp-postcss";
+import cssnano from "cssnano";
+import mqpacker from "css-mqpacker";
+import autoprefixer from "autoprefixer";
 import sourcemaps from "gulp-sourcemaps";
 import plumber from "gulp-plumber";
 import browsersync from "browser-sync";
@@ -17,33 +18,18 @@ import yargs from "yargs";
 const argv = yargs.argv,
   production = !!argv.production;
 
+const processors = [
+  cssnano,
+  mqpacker,
+  autoprefixer
+];
+
 gulp.task("styles", () => {
   return gulp.src(paths.styles.src)
     .pipe(gulpif(!production, sourcemaps.init()))
     .pipe(plumber())
     .pipe(sass())
-    .pipe(groupmedia())
-    .pipe(gulpif(production, autoprefixer({
-      cascade: false,
-      grid: true
-    })))
-    .pipe(gulpif(production, mincss({
-      compatibility: "ie8", level: {
-        1: {
-          specialComments: 0,
-          removeEmpty: true,
-          removeWhitespace: true
-        },
-        2: {
-          mergeMedia: true,
-          removeEmpty: true,
-          removeDuplicateFontRules: true,
-          removeDuplicateMediaBlocks: true,
-          removeDuplicateRules: true,
-          removeUnusedAtRules: false
-        }
-      }
-    })))
+    .pipe(gulpif(production, postcss(processors)))
     .pipe(gulpif(production, rename({
       suffix: ".min"
     })))
