@@ -1,23 +1,28 @@
-"use strict";
+'use strict';
 
-import { paths } from "../gulpfile.babel";
-import gulp from "gulp";
-import pug from "gulp-pug";
-import gulpif from "gulp-if";
-import replace from "gulp-replace";
-import browsersync from "browser-sync";
-import yargs from "yargs";
+import { paths } from '../gulpfile.babel';
+import gulp from 'gulp';
+import pug from 'gulp-pug';
+import typograf from 'gulp-typograf';
+import gulpif from 'gulp-if';
+import replace from 'gulp-replace';
+import browsersync from 'browser-sync';
+import yargs from 'yargs';
 
 const argv = yargs.argv,
-  production = !!argv.production;
+    production = !!argv.production;
 
-gulp.task("views", () => {
-  return gulp.src(paths.views.src)
-    .pipe(pug({
-      pretty: true
-    }))
-    .pipe(gulpif(production, replace(".css", ".min.css")))
-    .pipe(gulpif(production, replace(".js", ".min.js")))
-    .pipe(gulp.dest(paths.views.dist))
-    .pipe(browsersync.stream());
+gulp.task('views', () => {
+    return gulp
+        .src(paths.views.src)
+        .pipe(
+            pug({
+                pretty: true,
+            })
+        )
+        .pipe(typograf({ locale: ['ru', 'en-US'] }))
+        .pipe(gulpif(production, replace('.css', '.min.css')))
+        .pipe(gulpif(production, replace('.js', '.min.js')))
+        .pipe(gulp.dest(paths.views.dist))
+        .pipe(browsersync.stream());
 });
