@@ -1,7 +1,5 @@
-/* eslint-disable no-restricted-syntax */
 /* eslint-disable no-restricted-globals */
-import PerfectScrollbar from 'perfect-scrollbar';
-
+/* eslint-disable class-methods-use-this */
 class Header {
     constructor(element, dropdowns) {
         this.header = document.querySelector(element);
@@ -9,7 +7,7 @@ class Header {
     }
 
     toScrollFixHeader(header) {
-        window.addEventListener('scroll', function() {
+        window.addEventListener('scroll', () => {
             const scrolled = pageYOffset || document.body.scrollTop;
             if (scrolled > 0) {
                 header.classList.add('header--fixed');
@@ -19,62 +17,11 @@ class Header {
         });
     }
 
-    dropDownOpen(dropdowns) {
-        const $this = this;
-        for (const dropdown of dropdowns) {
-            const dropdownContainer = document.querySelector(dropdown);
-            const dropdownButton = dropdownContainer.querySelector(
-                '.dropdown-trigger'
-            );
-            const dropdownMenu = dropdownContainer.querySelector(
-                '.dropdown-menu'
-            );
-            dropdownButton.addEventListener('click', function(e) {
-                e.preventDefault();
-                if (dropdownMenu.classList.contains('dropdown-menu--active')) {
-                    dropdownButton.classList.remove('dropdown-trigger--active');
-                    dropdownMenu.classList.remove('dropdown-menu--active');
-                } else {
-                    $this.dropDownClose(dropdowns);
-                    dropdownButton.classList.add('dropdown-trigger--active');
-                    dropdownMenu.classList.add('dropdown-menu--active');
-                }
-            });
-        }
-    }
-
-    dropDownClose(dropdowns) {
-        for (const dropdown of dropdowns) {
-            const dropdownContainer = document.querySelector(dropdown);
-            const dropdownButton = dropdownContainer.querySelector(
-                '.dropdown-trigger'
-            );
-            const dropdownMenu = dropdownContainer.querySelector(
-                '.dropdown-menu'
-            );
-            dropdownButton.classList.remove('dropdown-trigger--active');
-            dropdownMenu.classList.remove('dropdown-menu--active');
-        }
-    }
-
-    perfectScroll(dropdownPhones) {
-        new PerfectScrollbar(dropdownPhones, {
-            wheelSpeed: 0.3,
-            wheelPropagation: false,
-        });
-    }
-
     init() {
         const $this = this;
         const { header } = $this;
-        const { dropdowns } = $this;
-        $this.dropDownOpen(dropdowns);
         $this.toScrollFixHeader(header);
-        $this.perfectScroll('.header-phone-all__wrapper');
     }
 }
 
-const header = new Header('.header', [
-    '.header__contacts',
-    '.header__buttons',
-]).init();
+new Header('.header', ['.header__contacts', '.header__buttons']).init();
