@@ -1,3 +1,4 @@
+/* eslint-disable import/extensions */
 /* eslint-disable import/no-unresolved */
 import '%modules%/header/header';
-import '%modules%/footer/footer';
+import '%modules%/popup/popup';
